Knopflerfish hangs the VM if it is not explicitly shut down.
In the case where Knopflerfish is embedded and the point of application
shutdown is not known to the embedding code, the
application never exits because of a non-daemon background thread.

To fix this the background thread created in /trunk/.../StartLevelImpl.java:93 
should be set to a daemon thread.

See also: http://sourceforge.net/tracker/index.php?func=detail&aid=1681705&group_id=82798&atid=567241


svn path=/dso/trunk/; revision=4716