CDV-586: Windows scripts broken for some paths. See also: http://forums.terracotta.org/forums/posts/list/731.page
The issue was that in Windows when JAVA_HOME is defined, and the path contains spaces, it will already have quotes around the path,
this was screwing up our scripts when we put yet another set of quotes around it before shortening it into the 8.3 format.

We now remove the quotes from JAVA_HOME before defined, and re-add it, before shortening to 8.3 format.

Also updated the copyright information.


svn path=/dso/trunk/; revision=7548