CDV-586: Windows scripts broken for some paths.
Applied broken path fixes for all other Windows scripts.
It was also broken for other constructed paths besides, JAVA_HOME.

NOTES:
The issue was that in Windows when JAVA_HOME is defined, and the path contains spaces, it will already have quotes around the path,
this was screwing up our scripts when we put yet another set of quotes around it before shortening it into the 8.3 format.

We now remove the quotes from JAVA_HOME if it is defined, and re-add it, before shortening to 8.3 format.

See also: revs 7548, rev7553, 7555, 7557 (trunk) 



svn path=/dso/trunk/; revision=7570