CDV-586: Windows scripts broken for some paths.
Reverting back to not using or converting paths to 8dot3 format. Windows facility for effecting this isn't very reliable - it may work on a certain machine, but wont on another with the same OS config. 

I've added an extra line however to remove the quotes from the incoming paths, and re-add it ourselves, which seems to be causing the issue the prompted this changeset.
The line was (and similar other lines):

   set JAVA_HOME="%JAVA_HOME:"=%"



svn path=/dso/trunk/; revision=7736